Pass the Play Application to the applicationModule function.
In some module I need to access the Play Configuration to get some
configured value. To achieve this the applicationModule now gets
passed the Play Application instance. I prefer to provide the Application
explicitely over accessing it via Play.current because it makes clear
that the Application is already available.
I chose to provide the Play Application instead of just the Configuration
because then Plugins are available as well.

The downside of this change is that it's not backward compatible.